adds menu class and extracts all menu related activity to this class from index and scoreboard
