+++ lib/game.js
@@ -110,7 +110,11 @@ class Game {
   }
 
   updateScore() {
-    this.score++
+    this.score++;
+  }
+
+  inProgress() {
+    return !this.snake.isDead();
   }
 
   getRandomCoords(){
