+++ lib/index.js
@@ -1,32 +1,24 @@
 require('./style.scss');
 
 var Game = require("./game");
-var MenuTyper = require("./menuTyper");
 var Menu = require("./menu");
-var ScoreBoard = require("./scoreBoard");
 
 var canvas = document.getElementById('canvas');
 var context = canvas.getContext('2d');
-var scoreBoard = new ScoreBoard();
-var menuTyper = new MenuTyper()
-var menu = new Menu(scoreBoard, menuTyper);
+var menu = new Menu();
 
 function setupPregameDisplay() {
   menu.setup();
-  $("#restartButton").on("click", function(){
-    gameLoop();
-  });
+  $("#restartButton").on("click", gameLoop);
 }
 
 function setupGameDisplay() {
   menu.prepareForGame();
-  $("#restartButton").hide();
-  $("#currentConversion").html("Binary:")
 }
 
 function setupPostgameDisplay(score) {
   $("#restartButton").show();
-  scoreBoard.getLeaderName(score);
+  menu.scoreBoard.getLeaderName(score);
 }
 
 function gameLoop() {
@@ -34,7 +26,7 @@ function gameLoop() {
   setupGameDisplay();
   game.setup();
   requestAnimationFrame(function gamePlay() {
-    if (!game.snake.isDead()) {
+    if (game.inProgress()) {
       game.gameLoop();
       setTimeout(function(){
         requestAnimationFrame(gamePlay);
