+++ lib/menu.js
@@ -1,7 +1,10 @@
+var MenuTyper = require("./menuTyper");
+var ScoreBoard = require("./scoreBoard");
+
 class Menu {
-  constructor(scoreBoard, menuTyper){
-    this.scoreBoard = scoreBoard;
-    this.menuTyper = menuTyper
+  constructor(){
+    this.scoreBoard = new ScoreBoard();
+    this.menuTyper = new MenuTyper();
   }
 
   setup() {
@@ -21,18 +24,20 @@ class Menu {
           this.showInstructions();
         }
         if ($(e.target).attr('target') === "scoreboard") {
-          this.scoreBoard.populate();
+          this.populateScoreboard(this.scoreBoard.currentHighScores);
         }
       }.bind(this));
   };
 
   scoreBoardListener() {
+    $(".add-leader").hide();
     $("#leaderName").on("keydown", function(event) {
       if (event.which == 13 || event.keycode == 13) {
         var leaderName = $("#leaderName").html();
         this.scoreBoard.updateHighScores(leaderName);
         $(".add-leader").hide();
         $("#leaderName").html("Enter Name Here.")
+        this.populateScoreboard(this.scoreBoard.currentHighScores)
       }
     }.bind(this));
   }
@@ -51,8 +56,20 @@ class Menu {
 
   prepareForGame(){
     this.clearStatusMenu();
+    $("#restartButton").hide();
+    $("#currentConversion").html("Binary:")
     $(".status p").remove();
     $("#status-tab").click();
+  }
+
+  populateScoreboard(highScores) {
+    $(".score").remove();
+    $("#scoreboard-post-type").hide();
+    highScores.forEach(function(score, index){
+      var scoreInfo = score.split("~")
+      $("#scoreboard-post-type").append("<p class='score'>" + index + ". " + scoreInfo[0] + " : " + scoreInfo[1] + "</p>" )
+    })
+    this.menuTyper.typeStuff(" cat leaderboard.txt", $("#scoreboard-typer"), "scoreboard");
   }
 
 }
