+++ lib/scoreBoard.js
@@ -1,6 +1,3 @@
-var MenuTyper = require("./menuTyper");
-var menuTyper = new MenuTyper()
-
 class ScoreBoard {
   constructor() {
     this.currentHighScores = this.getHighScores();
@@ -22,8 +19,6 @@ class ScoreBoard {
       })
 
       localStorage.setItem("highScores", this.currentHighScores);
-
-      this.populate();
   }
 
   getLeaderName(score){
@@ -62,16 +57,6 @@ class ScoreBoard {
 
   lowestScore() {
     return this.currentHighScores[this.numHighScores() - 1].split("~")[1];
-  }
-
-  populate() {
-    $(".score").remove();
-    $("#scoreboard-post-type").hide();
-    this.currentHighScores.forEach(function(score, index){
-      var scoreInfo = score.split("~")
-      $("#scoreboard-post-type").append("<p class='score'>" + index + ". " + scoreInfo[0] + " : " + scoreInfo[1] + "</p>" )
-    })
-    menuTyper.typeStuff(" cat leaderboard.txt", $("#scoreboard-typer"), "scoreboard");
   }
 
 }
